change res.status of deleteProduct()

+++ backend/controllers/product.controller.js
@@ -55,11 +55,18 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   // console.log("id:", id);
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Product not found." });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(204).json({ success: true, message: "Product deleted." });
   } catch (error) {
     console.error("Error in deleting product:", error);
-    res.status(404).json({ success: false, message: "Product not found." });
+    res.status(500).json({ success: false, message: "Server error." });
   }
 };
